+++ src/Objects/Item.java
@@ -70,7 +70,7 @@ public class Item {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
